[Minor] Inquiry block location to master before checking local store
Block location responses form the master has overloaded meaning.
It does not only resolve the location, but also guarantees the block is
committed. This assumption is needed when the block consumer is placed
on the executor where the block producer resides. Early resolving via
local block store spoils this semantic, because BlockManagerWorker does
not track the status of each block.

This commit makes BlockManagerWorker to inquiry the location of the
specified block before resolving the block via local block store.